Extract pagination helper into central module
This will be used by all transports so there is no reason to duplicate code
